requirements.txt
fastapi
uvicorn
pydantic
google-genai
aiofiles
python-multipart
python-dotenv
